dird: extracting `IsSameStorage()`
